Separated the token requests in Token.py
A problem of circular import was detected,
and putting aside token requests solved this issue
